fix: calculate spectral diff to absolute reference
Amplitude is default calculated to ref=np.max what means that it first
shifts all to its maximum before calculating. But when a low level sig-
nal is compared to a high level signal this itself is a difference that
should be perceived. So set the reference to maximum of 1.0 and no
shifting is done.

Signed-off-by: Conrad Zelck <git@simpel.cc>