feat: introduce option --scan_size
It's the size of the fft window and correlation window in samples.
Increasing the scan_size makes the tool faster but the threshold had to
be lowered as it is more insensitive.

Example: scan_size of 12000 (default 512) speeds up. A file of 3h18min
scans about 2 minutes on default values but needs only 15 seconds with
the scans_size of 12000 samples. It's 1/8 of the default time.
But it gives only nearly the same results on threshold 50 (default 60).

Signed-off-by: Conrad Zelck <git@simpel.cc>